Move ServletContainer to WebServlet
