+++ rest/jersey2/src/main/java/org/seedstack/seed/rest/jersey2/internal/SeedApplication.java
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.seedstack.seed.rest.jersey2;
+package org.seedstack.seed.rest.jersey2.internal;
 
 import org.glassfish.jersey.server.ResourceConfig;
