+++ rest/jersey2/src/main/java/org/seedstack/seed/rest/jersey2/internal/SeedServletContainer.java
@@ -1,0 +1,12 @@
+package org.seedstack.seed.rest.jersey2.internal;
+
+import org.glassfish.jersey.servlet.ServletContainer;
+import org.seedstack.seed.web.WebServlet;
+
+@WebServlet("/*")
+public class SeedServletContainer extends ServletContainer {
+
+    public SeedServletContainer() {
+        super(new SeedApplication());
+    }
+}
